requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
Automat==25.4.16
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blockbuster==1.5.24
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
constantly==23.10.4
cryptography==44.0.3
cssselect==1.3.0
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
elastic-transport==8.17.1
elasticsearch==8.14.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
fake-useragent==2.2.0
Faker==37.4.2
fastapi==0.116.1
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
forbiddenfruit==0.1.4
frozenlist==1.7.0
functions==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
googlemaps==4.10.0
greenlet==3.2.3
groq==0.26.0
grpcio==1.73.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.75
langchain-google-genai==2.0.10
langchain-groq==0.3.2
langchain-openai==0.3.17
langchain-postgres==0.0.15
langchain-text-splitters==0.3.8
langgraph==0.6.6
langgraph-api==0.2.48
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
langgraph-cli==0.3.1
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.2.1
langgraph-sdk==0.2.4
langsmith==0.3.45
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.3.0
openai==1.86.0
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.2
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.6
pickleshare==0.7.5
pillow==11.2.1
pip-tools==7.4.1
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.8.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
schemas==0.7.1
scrapeops_scrapy==0.5.6
scrapeops_scrapy_proxy_sdk==1.0
Scrapy==2.13.3
scrapy-fake-useragent==1.4.4
scrapy-playwright==0.0.43
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
selenium==4.33.0
sentry-sdk==2.35.2
serpapi==0.1.5
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.0
structlog==25.4.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tld==0.13.1
tldextract==5.3.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
truststore==0.10.1
Twisted==25.5.0
typer==0.16.0
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.4.0
user-agents==2.2.0
uvicorn==0.34.3
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.0.5
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
python-docx==1.1.2
